Extract the UI lib from popup - this makes updating state way simpler, without needing to muck about in the component tree. Just update the internal state, and rerender.
