+++ lua/neogit/popups/branch_config/actions.lua
@@ -41,7 +41,7 @@ function M.update_pull_rebase()
 end
 
 function M.merge_config(branch)
-  return a.void(function(popup, c)
+  local fn = function()
     local target = FuzzyFinderBuffer.new(git.refs.list_branches()):open_async { prompt_prefix = "upstream" }
     if not target then
       return
@@ -60,13 +60,14 @@ function M.merge_config(branch)
     git.config.set("branch." .. branch .. ".merge", merge_value)
     git.config.set("branch." .. branch .. ".remote", remote_value)
 
-    c.value = merge_value
-    popup:repaint_config()
-  end)
+    return merge_value
+  end
+
+  return a.wrap(fn, 2)
 end
 
 function M.description_config(branch)
-  return a.void(function(popup, c)
+  local fn = function()
     client.wrap(git.cli.branch.edit_description, {
       autocmd = "NeogitDescriptionComplete",
       msg = {
@@ -74,9 +75,10 @@ function M.description_config(branch)
       },
     })
 
-    c.value = git.config.get("branch." .. branch .. ".description"):read()
-    popup:repaint_config()
-  end)
+    return git.config.get("branch." .. branch .. ".description"):read()
+  end
+
+  return a.wrap(fn, 2)
 end
 
 return M
